升级instep到2.0.3，防止发号器因为不恰当的参数导致错误行为。

+++ framework/src/main/java/me/insidezhou/southernquiet/util/SnowflakeIdGenerator.java
@@ -81,7 +81,7 @@ public class SnowflakeIdGenerator extends LongIdGenerator implements IdGenerator
     }
 
     @Override
-    public long getSequenceFromId(long id) {
-        return (id << 64 - getSequenceBits()) >>> (64 - getSequenceBits());
+    public int getSequenceFromId(long id) {
+        return (int) (id << 64 - getSequenceBits()) >>> (64 - getSequenceBits());
     }
 }
